concord-server: remove deprecated SecretEntry
